Functional Item but Error in Custom Tab
Item and "stolen" item texture from google images. Custom Tab is made after cadiboos tutorial but i was very tired when i made it. so if anyone could look at the Link to the tutorial in dc or fix the tab and explain how, i would very much appreciate it